+++ src/Misc.hpp
@@ -6,6 +6,6 @@ Suit charToSuit(char stringSuit);
 std::string suitToString(Suit suit);
 std::string positionToString(Position position);
 void generateScoringChart();
-uint8_t randomUint8(int min, int max);
+uint8_t randomUint8(uint8_t min, uint8_t max);
 
 #endif
